// src/components/molecules/datepicker/enums/datepicker.component.enumerations.ts
enum DatepickerComponentBorderEnum {
  NONE = 'none',
  SOLID = 'solid',
}

enum DatepickerComponentBorderRadiusEnum {
  NONE = 'none',
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}

enum DatepickerComponentBoxShadowEnum {
  NONE = 'none',
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}

export {
  DatepickerComponentBorderEnum,
  DatepickerComponentBorderRadiusEnum,
  DatepickerComponentBoxShadowEnum,
};